projeler/Blog_Yazici/requirements.txt
requests==2.32.5
Pillow==12.1.1
opencv-python-headless==4.10.0.84
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
httplib2==0.22.0
google-api-core==2.24.2
